feat: add intro text animation
