+++ styles/globals.css
@@ -35,6 +35,6 @@
 body,
 body > div {
     min-height: 100vh;
-    background: #222222;
+    background: #e2e2e2;
     margin: 0;
 }
